+++ uninstall.php
@@ -24,10 +24,13 @@ function wimbblock_uninstall_delete_options() {
 	foreach ( $wimbblock_option_names as $key => $value ) {
 		delete_option( $value['option_name'] );
 	}
-	delete_transient( 'wimbblock_emergency_stop' );
+	delete_transient( 'wimbblock_always' );
+	delete_transient( 'wimbblock_browsers' );
 	delete_transient( 'wimbblock_crawlers' );
+	delete_transient( 'wimbblock_emergency_stop' );
 	delete_transient( 'wimbblock_logfile' );
 	delete_transient( 'wimbblock_logging_levels' );
+	delete_transient( 'wimbblock_systems' );
 }
 
 global $wpdb;