+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 260901 /
+
+* fixes errors in handling of transients
+* bumped minimum versions of agents: Chrome, Edge and Chrome based browsers < 147, Samsung Internet Browser < 30,
+
 ### 260823 / 2.8.1
 
 * changed testing and creating robots.txt